+++ UVOCBot/Services/WelcomeMessageService.cs
@@ -1,10 +1,12 @@
 ﻿using FuzzySharp;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Remora.Discord.API.Abstractions.Gateway.Events;
 using Remora.Discord.API.Abstractions.Objects;
 using Remora.Discord.API.Abstractions.Rest;
 using Remora.Discord.API.Objects;
 using Remora.Discord.Commands.Contexts;
+using Remora.Discord.Commands.Services;
 using Remora.Discord.Core;
 using Remora.Results;
 using System;
@@ -22,33 +24,33 @@ namespace UVOCBot.Services
     public class WelcomeMessageService : IWelcomeMessageService
     {
         private readonly ILogger<WelcomeMessageService> _logger;
-        private readonly InteractionContext _context;
         private readonly ICensusApiService _censusApi;
         private readonly IDbApiService _dbApi;
         private readonly IDiscordRestChannelAPI _channelApi;
         private readonly IDiscordRestGuildAPI _guildApi;
-        private readonly IReplyService _responder;
+        private readonly IServiceProvider _serviceProvider;
 
         public WelcomeMessageService(
             ILogger<WelcomeMessageService> logger,
-            InteractionContext context,
             ICensusApiService censusApi,
             IDbApiService dbApi,
             IDiscordRestChannelAPI channelApi,
             IDiscordRestGuildAPI guildApi,
-            IReplyService responder)
+            IServiceProvider serviceProvider)
         {
             _logger = logger;
-            _context = context;
             _censusApi = censusApi;
             _dbApi = dbApi;
             _channelApi = channelApi;
             _guildApi = guildApi;
-            _responder = responder;
+            _serviceProvider = serviceProvider;
         }
 
         public async Task<Result> SendWelcomeMessage(IGuildMemberAdd gatewayEvent, CancellationToken ct = default)
         {
+            if (!gatewayEvent.User.HasValue || gatewayEvent.User.Value is null)
+                return Result.FromSuccess();
+
             // Get the welcome message settings
             Result<GuildWelcomeMessageDto> welcomeMessageResult = await GetGuildWelcomeMessage(gatewayEvent.GuildID.Value, ct).ConfigureAwait(false);
             if (!welcomeMessageResult.IsSuccess)
@@ -61,11 +63,11 @@ namespace UVOCBot.Services
             // Make some nickname guesses
             IEnumerable<string>? nicknameGuesses = null;
             if (welcomeMessage.DoIngameNameGuess)
-                nicknameGuesses = await DoFuzzyNicknameGuess(_context.User.Username, welcomeMessage.OutfitId, ct).ConfigureAwait(false);
+                nicknameGuesses = await DoFuzzyNicknameGuess(gatewayEvent.User.Value.Username, welcomeMessage.OutfitId, ct).ConfigureAwait(false);
 
             // Prepare components of the welcome message
-            List<ButtonComponent> messageButtons = CreateWelcomeMessageButtons(welcomeMessage, nicknameGuesses, _context.User.ID.Value);
-            string messageContent = SubstituteMessageVariables(welcomeMessage.Message);
+            List<ButtonComponent> messageButtons = CreateWelcomeMessageButtons(welcomeMessage, nicknameGuesses, gatewayEvent.User.Value.ID.Value);
+            string messageContent = SubstituteMessageVariables(welcomeMessage.Message, gatewayEvent.User.Value.ID);
 
             // Send the welcome message
             Result<IMessage> sendWelcomeMessageResult = await _channelApi.CreateMessageAsync(
@@ -78,7 +80,7 @@ namespace UVOCBot.Services
             // Assign default roles
             await ModifyRoles(
                 gatewayEvent.GuildID,
-                _context.User.ID,
+                gatewayEvent.User.Value.ID,
                 gatewayEvent.Roles.ToList(),
                 rolesToAdd: welcomeMessage.DefaultRoles,
                 ct: ct).ConfigureAwait(false);
@@ -130,9 +132,9 @@ namespace UVOCBot.Services
             return messageButtons;
         }
 
-        private string SubstituteMessageVariables(string welcomeMessage)
+        private static string SubstituteMessageVariables(string welcomeMessage, Snowflake userId)
         {
-            return welcomeMessage.Replace("<name>", Formatter.UserMention(_context.User.ID));
+            return welcomeMessage.Replace("<name>", Formatter.UserMention(userId));
         }
 
         #endregion
@@ -147,20 +149,24 @@ namespace UVOCBot.Services
             if (gatewayEvent.Data.Value is null || gatewayEvent.Data.Value.CustomID.Value is null)
                 return Result.FromSuccess();
 
+            IUser? user = gatewayEvent.GetUser();
+            if (user is null)
+                return Result.FromSuccess();
+
             ComponentIdFormatter.Parse(gatewayEvent.Data.Value.CustomID.Value, out ComponentAction _, out string payload);
             string[] payloadComponents = payload.Split('@');
             ulong userId = ulong.Parse(payloadComponents[0]);
 
             // Check that the user who clicked the button is the focus of the welcome message
-            if (_context.User.ID.Value != userId)
+            if (user.ID.Value != userId)
             {
-                await _responder.RespondWithSuccessAsync("Hold it, bud. You can't do that!", ct, new AllowedMentions()).ConfigureAwait(false);
+                await GetReplyService(gatewayEvent).RespondWithSuccessAsync("Hold it, bud. You can't do that!", ct, new AllowedMentions()).ConfigureAwait(false);
                 return Result.FromSuccess();
             }
 
-            await _guildApi.ModifyGuildMemberAsync(gatewayEvent.GuildID.Value, _context.User.ID, payloadComponents[1], ct: ct).ConfigureAwait(false);
+            await _guildApi.ModifyGuildMemberAsync(gatewayEvent.GuildID.Value, user.ID, payloadComponents[1], ct: ct).ConfigureAwait(false);
 
-            Result<IMessage> alertResponse = await _responder.RespondWithSuccessAsync(
+            Result<IMessage> alertResponse = await GetReplyService(gatewayEvent).RespondWithSuccessAsync(
                 $"Your nickname has been updated to { Formatter.Bold(payloadComponents[1]) }!",
                 ct).ConfigureAwait(false);
 
@@ -171,7 +177,7 @@ namespace UVOCBot.Services
 
         public async Task<Result> InformNicknameNoMatch(IInteractionCreate gatewayEvent, CancellationToken ct = default)
         {
-            Result<IMessage> alertResponse = await _responder.RespondWithSuccessAsync(
+            Result<IMessage> alertResponse = await GetReplyService(gatewayEvent).RespondWithSuccessAsync(
                 "Please set your nickname to the name of your PlanetSide 2 character!",
                 ct).ConfigureAwait(false);
 
@@ -235,6 +241,10 @@ namespace UVOCBot.Services
             if (gatewayEvent.Data.Value is null || gatewayEvent.Data.Value.CustomID.Value is null)
                 return Result.FromSuccess();
 
+            IUser? user = gatewayEvent.GetUser();
+            if (user is null)
+                return Result.FromSuccess();
+
             // Get the welcome message settings
             Result<GuildWelcomeMessageDto> welcomeMessage = await GetGuildWelcomeMessage(gatewayEvent.GuildID.Value.Value, ct).ConfigureAwait(false);
             if (!welcomeMessage.IsSuccess)
@@ -244,16 +254,16 @@ namespace UVOCBot.Services
             ulong userId = ulong.Parse(payload);
 
             // Check that the user who clicked the button is the focus of the welcome message
-            if (_context.User.ID.Value != userId)
+            if (user.ID.Value != userId)
             {
-                await _responder.RespondWithSuccessAsync("Hold it, bud. You can't do that!", ct, new AllowedMentions()).ConfigureAwait(false);
+                await GetReplyService(gatewayEvent).RespondWithSuccessAsync("Hold it, bud. You can't do that!", ct, new AllowedMentions()).ConfigureAwait(false);
                 return Result.FromSuccess();
             }
 
             // Remove the default roles and add the alternate roles
             Result roleChangeResult = await ModifyRoles(
                 gatewayEvent.GuildID.Value,
-                _context.User.ID,
+                user.ID,
                 gatewayEvent.Member.Value?.Roles.ToList(),
                 welcomeMessage.Entity.AlternateRoles,
                 welcomeMessage.Entity.DefaultRoles,
@@ -267,7 +277,7 @@ namespace UVOCBot.Services
 
             // Inform the user of their role change
             string rolesStringList = string.Join(' ', welcomeMessage.Entity.AlternateRoles.Select(r => Formatter.RoleMention(r)));
-            Result<IMessage> alertResponse = await _responder.RespondWithSuccessAsync(
+            Result<IMessage> alertResponse = await GetReplyService(gatewayEvent).RespondWithSuccessAsync(
                 $"You've been given the following roles: { rolesStringList }",
                 ct,
                 new AllowedMentions()).ConfigureAwait(false);
@@ -322,5 +332,16 @@ namespace UVOCBot.Services
         }
 
         #endregion
+
+        private IReplyService GetReplyService(IInteractionCreate gatewayEvent)
+        {
+            Result<InteractionContext> context = gatewayEvent.ToInteractionContext();
+            if (!context.IsSuccess)
+                throw new Exception("Failed to get interaction context from gateway event");
+
+            IServiceScope scope = _serviceProvider.CreateScope();
+            scope.ServiceProvider.GetRequiredService<ContextInjectionService>().Context = context.Entity;
+            return scope.ServiceProvider.GetRequiredService<IReplyService>();
+        }
     }
 }